support markdown inclusions
