+++ src/genomehubs-ui/src/client/views/components/Markdown.jsx
@@ -1,34 +1,188 @@
-import React, { useEffect } from "react";
-import {
-  RehypeComponentsList,
-  compile,
-  processProps,
-} from "./MarkdownFunctions";
+import React, { createElement, useEffect } from "react";
+import { basename, siteName } from "../reducers/location";
 
-import MarkdownInclude from "./MarkdownInclude";
+import AggregationIcon from "./AggregationIcon";
+import Grid from "@material-ui/core/Grid";
+import Highlight from "./Highlight";
+import NavLink from "./NavLink";
+import Report from "./Report";
+import Template from "./Template";
+import Toggle from "./Toggle";
+import Tooltip from "@material-ui/core/Tooltip";
 import classnames from "classnames";
 import { compose } from "recompose";
+import gfm from "remark-gfm";
+import { h } from "hastscript";
+import rehypeRaw from "rehype-raw";
+import rehypeReact from "rehype-react";
+import remarkDirective from "remark-directive";
+import remarkParse from "remark-parse";
+import remarkReact from "remark-react";
+import remarkRehype from "remark-rehype";
 import styles from "./Styles.scss";
+import unified from "unified";
+import { visit } from "unist-util-visit";
 import withPages from "../hocs/withPages";
 import { withStyles } from "@material-ui/core/styles";
 
+const pagesUrl = PAGES_URL;
+const webpackHash = __webpack_hash__ || COMMIT_HASH;
+// const webpackHash = COMMIT_HASH;
+
+export const processProps = (props, newProps = {}) => {
+  for (const [key, value] of Object.entries(props)) {
+    if (value === false) {
+      newProps[key] = value;
+    } else if (value == "") {
+      newProps[key] = true;
+    } else if (key == "className") {
+      newProps["className"] = styles[value];
+    } else if (key.startsWith("exclude")) {
+      newProps[key] = value.split(",");
+    } else if (key == "src") {
+      if (PAGES_URL.startsWith("http")) {
+        newProps["src"] = `${pagesUrl}${value.replace(/^\/static/, "")}`;
+      } else {
+        newProps["src"] = value.replace(
+          /^\/static\//,
+          `${basename}/static/${webpackHash}/`
+        );
+      }
+    } else if (key == "xs") {
+      newProps["xs"] = value * 1;
+    } else if (key == "spacing") {
+      newProps["spacing"] = value * 1;
+    } else if (key != "pageId") {
+      newProps[key] = value;
+    }
+  }
+  return newProps;
+};
+
+export const RehypeComponentsList = {
+  a: (props) => <NavLink {...processProps(props)} />,
+  aggregation: (props) => <AggregationIcon method={props.method} />,
+  grid: (props) => {
+    let { toggle, expand, title, ...gridProps } = props;
+    if (toggle && toggle !== true && toggle !== "true") {
+      toggle = false;
+    }
+    if (props.hasOwnProperty("toggle")) {
+      return (
+        <Toggle {...processProps({ toggle, expand, title })}>
+          <Grid {...processProps(gridProps)} />
+        </Toggle>
+      );
+    } else {
+      return <Grid {...processProps(props)} />;
+    }
+  },
+  hub: (props) => <span {...processProps(props)}>{siteName}</span>,
+  img: (props) => (
+    <div className={styles.centerContent}>
+      <img {...processProps(props)} alt={props.alt.toString()} />
+    </div>
+  ),
+  include: (props) => {
+    let nested = <Nested pgId={props.pageId} />;
+    return (
+      <Grid
+        {...processProps(props)}
+        item
+        xs={8}
+        className={styles.reportContainer}
+      >
+        {nested}
+      </Grid>
+    );
+  },
+  item: (props) => (
+    <Grid {...processProps(props)} item className={styles.reportContainer} />
+  ),
+  markdown: (props) => <Nested pgId={props.pageId} />,
+  pre: (props) => <Highlight {...props} />,
+  report: (props) => (
+    <Report {...processProps(props)} className={styles.reportContainer} />
+  ),
+  span: (props) => <span {...processProps(props)} />,
+  templat: (props) => (
+    <Template {...processProps(props)} className={styles.reportContainer} />
+  ),
+  tooltip: (props) => {
+    return (
+      <Tooltip {...processProps(props, { placement: "top" })}>
+        <span>{props.children}</span>
+      </Tooltip>
+    );
+  },
+};
+
+export function compile(val, components = RehypeComponentsList) {
+  const processor = unified()
+    .use(remarkParse)
+    .use(remarkReact)
+    .use(gfm)
+    .use(remarkDirective)
+    .use(htmlDirectives)
+    .use(remarkRehype)
+    .use(rehypeRaw)
+    .use(rehypeReact, {
+      createElement,
+      components,
+    });
+
+  const ast = processor.runSync(processor.parse(val));
+
+  return {
+    ast,
+    contents: processor.stringify(ast),
+  };
+}
+
+export function htmlDirectives() {
+  return transform;
+
+  function transform(tree) {
+    let index = { i: 0 };
+    visit(
+      tree,
+      ["textDirective", "leafDirective", "containerDirective"],
+      (node) => ondirective(node, index)
+    );
+  }
+
+  function ondirective(node, index) {
+    let data = node.data || (node.data = {});
+    let hast = h(node.name, node.attributes);
+    data.hName = hast.tagName;
+    data.hProperties = hast.properties;
+    if (data.hName == "report") {
+      data.hProperties.id = `report-${data.hProperties.report}-${index.i}`;
+      index.i += 1;
+    }
+  }
+}
+
 const Markdown = ({
   classes,
+  pgId,
   pageId,
-  pagesById,
+  pages,
   fetchPages,
   siteStyles,
   components = {},
 }) => {
   useEffect(() => {
-    if (pageId && !pagesById) {
+    if (pgId && !pages.byId[pgId]) {
+      fetchPages(pgId);
+    }
+    if (pageId && !pages.byId[pgId]) {
       fetchPages(pageId);
     }
-  }, [pageId]);
+  }, [pgId, pageId]);
 
-  const { contents, ast } = compile(pagesById, {
+  const { contents, ast } = compile(pages.byId[pgId || pageId], {
     ...RehypeComponentsList,
-    include: (props) => <MarkdownInclude {...processProps(props)} />,
     ...components,
   });
   let css;
@@ -39,5 +193,7 @@ const Markdown = ({
   }
   return <div className={css}>{contents}</div>;
 };
+
+const Nested = compose(withPages, withStyles(styles))(Markdown);
 
 export default compose(withPages, withStyles(styles))(Markdown);
